Consolidate Eagle test into existing Mistral Large 3 nightly test
Move the Mistral Large 3 Eagle test into the same test file as the base
Mistral Large 3 test, removing the need for a separate workflow step.
This keeps the Eagle test within the existing nightly test job.
