+++ .github/workflows/nightly-test-nvidia.yml
@@ -432,6 +432,7 @@ jobs:
           SGLANG_ENABLE_JIT_DEEPGEMM: "0"
         run: |
           rm -rf test/performance_profiles_mistral_large3/
+          rm -rf test/performance_profiles_mistral_large3_eagle/
           cd test
           IS_BLACKWELL=1 python3 nightly/test_mistral_large3_perf.py
 
@@ -442,26 +443,6 @@ jobs:
           GITHUB_RUN_NUMBER: ${{ github.run_number }}
         run: |
           python3 scripts/ci/publish_traces.py --traces-dir test/performance_profiles_mistral_large3
-
-      - name: Run Mistral-Large-3-Eagle nightly performance test
-        if: always()
-        timeout-minutes: 180
-        env:
-          TRACE_BASE_URL: https://raw.githubusercontent.com/sglang-bot/sglang-ci-data/main/traces/${{ github.run_id }}
-          PERFETTO_RELAY_URL: ${{ vars.PERFETTO_RELAY_URL }}
-          GPU_CONFIG: "8-gpu-b200"
-          SGLANG_ENABLE_JIT_DEEPGEMM: "0"
-        run: |
-          rm -rf test/performance_profiles_mistral_large3_eagle/
-          cd test
-          IS_BLACKWELL=1 python3 nightly/test_mistral_large3_eagle_perf.py
-
-      - name: Publish Mistral-Large-3-Eagle traces to storage repo
-        env:
-          GITHUB_TOKEN: ${{ secrets.GH_PAT_FOR_NIGHTLY_CI_DATA }}
-          GITHUB_RUN_ID: ${{ github.run_id }}
-          GITHUB_RUN_NUMBER: ${{ github.run_number }}
-        run: |
           python3 scripts/ci/publish_traces.py --traces-dir test/performance_profiles_mistral_large3_eagle
 
       - name: Run DeepSeek v3.1 nightly performance test
